+++ erde/constants.lua
@@ -25,10 +25,6 @@ C.DEBUG = false
   -- User specified library to use for bit operations.
 C.BITLIB = nil
 
--- Flag to capture and rewrite errors produced from compiled code. Assumes that
--- Erde is in LUA_PATH.
-C.REWRITE_ERRORS = true
-
 -- -----------------------------------------------------------------------------
 -- Lua Target
 -- -----------------------------------------------------------------------------